+++ src/background.js
@@ -17,22 +17,49 @@ function openInPopup(url) {
  * Share Current Tab
  * @return {void}       -
  */
-function shareCurrentTab() {
-  browser.tabs.query({
-    active: true,
-    currentWindow: true,
-  }).then((tabs) => {
-    const url = tabs[0].url;
-    const title = tabs[0].title || url;
-    browser.storage.local.get('url').then((storage) => {
-      if(storage.url){
-        const shareUrl = `${storage.url}?post=${encodeURIComponent(url)}&title=${encodeURIComponent(title)}&source=bookmarklet`;
-        openInPopup(shareUrl);
-      }else{
-        browser.runtime.openOptionsPage()
-      }
-    });
+function shareCurrentTab(tab) {
+  const url = tab.url;
+  const title = tab.title || url;
+  browser.storage.local.get().then((storage) => {
+    if(!storage.url){
+      browser.runtime.openOptionsPage();
+      return;
+    }
+    const shareUrl = `${storage.url}?post=${encodeURIComponent(url)}&title=${encodeURIComponent(title)}&source=bookmarklet`;
+    openInPopup(shareUrl);
   });
 }
 
+/**
+ * Show Page Action on current page if activated
+ * @param  {string} tabId - Current tab id
+ * @return {void}         -
+ */
+function showPageAction(tabId){
+  browser.storage.local.get().then((storage) => {
+    storage.pageAction ? browser.pageAction.show(tabId) : browser.pageAction.hide(tabId);
+  });
+}
+
+/**
+ * Show page action with given active information
+ * @param  {object} activeInfo  -
+ * @return {void}               -
+ */
+function tabActivation(activeInfo){
+  showPageAction(activeInfo.tabId);
+}
+
+/*
+Initialize
+ */
+browser.tabs.query({active: true, currentWindow: true}).then((tabs) => {
+  showPageAction(tabs[0].id);
+});
+
+/*
+Listeners
+ */
+browser.tabs.onActivated.addListener(tabActivation);
 browser.browserAction.onClicked.addListener(shareCurrentTab);
+browser.pageAction.onClicked.addListener(shareCurrentTab);